backend/requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
djangorestframework-gis>=1.2.0
django-filter>=25.2
psycopg2-binary>=2.9.11
celery>=5.5.3
redis>=7.0.1
django-cors-headers>=4.9.0
python-dotenv>=1.2.1
gunicorn>=23.0.0